requirements.txt
asgiref==3.9.0
Django==5.2.4
django-csp==4.0
django-filter==25.1
django-taggit==6.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.94.2
packaging==25.0
pillow==11.3.0
PyJWT==2.10.1
PyMySQL==1.1.1
setuptools==80.9.0
simplejwt==2.0.1
sqlparse==0.5.3
typing==3.7.4.3
wheel==0.45.1
